test(counters_stable/wasm): view counters > counts
