+++ examples/counters_stable/tests/web/fixtures/mod.rs
@@ -1,0 +1,1 @@
+pub mod counters_page;
